divided the AddDrink page into three components, separated the styles
